ToraOps/requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
cffi==1.14.4
cryptography==3.3.1
Django==3.1.5
django-bootstrap==0.2.4
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
numpy==1.19.5
pandas==1.2.0
paramiko==2.7.2
Pillow==8.1.0
pycparser==2.20
PyJWT==2.0.0
PyMySQL==1.0.2
PyNaCl==1.4.0
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
django-grappelli==2.15.1
django-apscheduler==0.6.0
django-multiselectfield==0.1.12
coreapi==2.3.3
ansible==4.4.0
pymssql==2.2.2